requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
py-cord==2.4.1
python-dotenv==1.0.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
